refactor grpc server, servicer focus on grpc protocol
